requirements.txt
websocket-client
tornado
flask
requests
numba